// src/models/Condition.ts
export enum Condition {
    BLINDED = "blinded",
    CHARMED = "charmed",
    DEAFENED = "deafened",
    EXHAUSTION = "exhaustion",
    FRIGHTENED = "frightened",
    PARALYZED = "paralyzed",
    PETRIFIED = "petrified",
    POISONED = "poisoned",
    PRONE = "prone",
    GRAPPLED = "grappled",
    RESTRAINED = "restrained",
    STUNNED = "stunned",
    UNCONSCIOUS = "unconscious"
}