// src/models/DamageType.ts
export enum DamageType {
    POISON = "poison",
    PSYCHIC = "psychic",
    FIRE = "fire",
    LIGHTNING = "lightning",
    BLUDGEONING = "bludgeoning",
    PIERCING = "piercing",
    SLASHING = "slashing",
    COLD = "cold",
    NECROTIC = "necrotic",
    THUNDER = "thunder",
    ACID = "acid",
    FORCE = "force",
    RADIANT = "radiant"
}